requirements.txt
asgiref==3.6.0
Django==4.2.1
django-cors-headers==4.0.0
django-taggit==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
numpy==1.24.3
pandas==2.0.1
Pillow==9.5.0
PyJWT==2.7.0
python-dateutil==2.8.2
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
